+++ app/forms.py
@@ -6,6 +6,7 @@ All classes derived from ``FlaskForm`` imported from ``Flask-WTF``.
 ``Flask-WTF`` is a ``Flask`` extension built on top of ``WTForms``.
 """
 from flask import current_app
+from flask_babel import lazy_gettext as _l
 from flask_pagedown.fields import PageDownField
 from flask_wtf import FlaskForm
 from wtforms import (
@@ -30,16 +31,17 @@ class RegistrationForm(FlaskForm):
     """Form for registering a new user."""
 
     username = StringField(
-        "Username", validators=[DataRequired(), Length(max=64)]
+        _l("Username"), validators=[DataRequired(), Length(max=64)]
     )
     email = StringField(
-        "Email", validators=[DataRequired(), Email(), Length(max=120)]
+        _l("Email"), validators=[DataRequired(), Email(), Length(max=120)]
     )
-    password = PasswordField("Password", validators=[DataRequired()])
+    password = PasswordField(_l("Password"), validators=[DataRequired()])
     confirm_password = PasswordField(
-        "Confirm Password", validators=[DataRequired(), EqualTo("password")]
+        _l("Confirm Password"),
+        validators=[DataRequired(), EqualTo("password")],
     )
-    submit = SubmitField("Register")
+    submit = SubmitField(_l("Register"))
 
     # noinspection PyMethodMayBeStatic
     def validate_username(self, username: StringField) -> None:
@@ -51,7 +53,7 @@ class RegistrationForm(FlaskForm):
         user = User.query.filter_by(username=username.data).first()
         reserved = username.data in current_app.config["RESERVED_USERNAMES"]
         if user is not None or reserved:
-            raise ValidationError("Username is taken")
+            raise ValidationError(_l("Username is taken"))
 
     # noinspection PyMethodMayBeStatic
     def validate_email(self, email: StringField) -> None:
@@ -63,64 +65,67 @@ class RegistrationForm(FlaskForm):
         user = User.query.filter_by(email=email.data).first()
         if user is not None:
             raise ValidationError(
-                "A user with this email address is already registered"
+                _l("A user with this email address is already registered")
             )
 
 
 class LoginForm(FlaskForm):
     """Form for logging in an existing user."""
 
-    username = StringField("Username", validators=[DataRequired()])
-    password = PasswordField("Password", validators=[DataRequired()])
-    remember_me = BooleanField("Remember Me")
-    submit = SubmitField("Sign In")
+    username = StringField(_l("Username"), validators=[DataRequired()])
+    password = PasswordField(_l("Password"), validators=[DataRequired()])
+    remember_me = BooleanField(_l("Remember Me"))
+    submit = SubmitField(_l("Sign In"))
 
 
 class PostForm(FlaskForm):
     """Form for creating new ``Post`` objects."""
 
-    title = StringField("Title", validators=[DataRequired()])
+    title = StringField(_l("Title"), validators=[DataRequired()])
     body = PageDownField(
-        "Body", validators=[DataRequired()], render_kw={"rows": 12}
+        _l("Body"), validators=[DataRequired()], render_kw={"rows": 12}
     )
-    submit = SubmitField("Submit")
+    submit = SubmitField(_l("Submit"))
 
 
 class ResetPasswordRequestForm(FlaskForm):
     """Form for requesting a reset email for an existing user."""
 
-    email = StringField("Email", validators=[DataRequired(), Email()])
-    submit = SubmitField("Request Password Reset")
+    email = StringField(_l("Email"), validators=[DataRequired(), Email()])
+    submit = SubmitField(_l("Request Password Reset"))
 
 
 class ResetPasswordForm(FlaskForm):
     """Form for resetting an existing user's password."""
 
-    password = PasswordField("Password", validators=[DataRequired()])
+    password = PasswordField(_l("Password"), validators=[DataRequired()])
     confirm_password = PasswordField(
-        "Confirm Password", validators=[DataRequired(), EqualTo("password")]
+        _l("Confirm Password"),
+        validators=[DataRequired(), EqualTo(_l("password"))],
     )
-    submit = SubmitField("Reset Password")
+    submit = SubmitField(_l("Reset Password"))
 
 
 class EditProfile(FlaskForm):
     """Form for editing an existing user's profile page."""
 
-    username = StringField("username", validators=[DataRequired()])
-    about_me = TextAreaField("About me", validators=[Length(min=0, max=140)])
-    submit = SubmitField("Submit")
+    username = StringField(_l("username"), validators=[DataRequired()])
+    about_me = TextAreaField(
+        _l("About me"), validators=[Length(min=0, max=140)]
+    )
+    submit = SubmitField(_l("Submit"))
 
 
 class EmptyForm(FlaskForm):
     """Empty form: Submit only."""
 
-    submit = SubmitField("Submit")
+    submit = SubmitField(_l("Submit"))
 
 
 class MessageForm(FlaskForm):
     """Form for creating new ``Message`` objects."""
 
     message = TextAreaField(
-        "Message", validators=[DataRequired(), Length(min=0, max=140)]
+        _l("Message"), validators=[DataRequired(), Length(min=0, max=140)]
     )
-    submit = SubmitField("Submit")
+    submit = SubmitField(_l("Submit"))